fix(transformer): added resolver build

+++ packages/graphql-transformer/src/transformer/GraphQLTransformer.ts
@@ -1,6 +1,7 @@
 import { writeFileSync } from "node:fs";
 import path from "node:path";
 import prettier from "@prettier/sync";
+import { buildSync } from "esbuild";
 import { TransformerContext } from "../context";
 import { IPluginFactory, TransformerPluginBase } from "../plugins";
 import { DocumentNode } from "../parser";
@@ -8,20 +9,39 @@ import { SchemaValidationError } from "../utils/errors";
 import { ensureOutputDirectory } from "../utils/output";
 import { TypeGenerator } from "../codegen";
 import { ResolverGenerator } from "../resolver/ResolverGenerator";
+import { ResolverConfig, DataSourceConfig } from "../utils/types";
+import { FieldResolver, FunctionResolver } from "../resolver";
 
 export interface GraphQLTransformerOptions {
   definition: string;
   plugins: IPluginFactory[];
-
-  mode: "development" | "production";
   // Absolute path for dev outputs
   outputDirectory: string;
+
+  mode: "development" | "production";
+  fieldResolvers?: Array<ResolverConfig>;
+  pipelineFunctions?: Array<ResolverConfig>;
+  dataSources?: Array<DataSourceConfig>;
 }
+
+export type FieldResolverOutput = {
+  typeName: string;
+  fieldName: string;
+  pipelineFunctions?: string[];
+  dataSource?: string;
+  code: string;
+};
+
+export type PipelineFunctionOutput = {
+  name: string;
+  dataSource?: string;
+  code: string;
+};
 
 export type TransformerOutput = {
   schema: string;
-  fieldResolvers: Record<string, unknown>;
-  pipelineFunctions: Record<string, unknown>;
+  fieldResolvers: Record<string, FieldResolverOutput>;
+  pipelineFunctions: Record<string, PipelineFunctionOutput>;
 };
 
 export class GraphQLTransformer {
@@ -57,7 +77,7 @@ export class GraphQLTransformer {
 
     // eslint-disable-next-line security/detect-non-literal-fs-filename
     writeFileSync(
-      path.resolve(outputPath, "schema-types.d.ts"),
+      path.resolve(outputPath, "schema-types.ts"),
       prettier.format(typesGen.generate(), { parser: "typescript" }),
       {
         encoding: "utf-8",
@@ -70,14 +90,16 @@ export class GraphQLTransformer {
     resolverGenerator.generate();
 
     for (const [name, resolver] of this.context.resolvers.entries()) {
-      // eslint-disable-next-line security/detect-non-literal-fs-filename
-      writeFileSync(
-        path.resolve(resolversDir, `${name}.ts`),
-        prettier.format(resolver.print(), { parser: "typescript" }),
-        {
-          encoding: "utf-8",
-        }
-      );
+      if (!resolver.isReadonly) {
+        // eslint-disable-next-line security/detect-non-literal-fs-filename
+        writeFileSync(
+          path.resolve(resolversDir, `${name}.ts`),
+          prettier.format(resolver.print(), { parser: "typescript" }),
+          { encoding: "utf-8" }
+        );
+
+        resolver.setSource(path.resolve(resolversDir, `${name}.ts`));
+      }
     }
   }
 
@@ -94,15 +116,67 @@ export class GraphQLTransformer {
 
     // Build resolvers
 
-    // for (const [name, resolver] of this.context.resolvers.entries()) {
-    //   if (resolver instanceof FunctionResolver) {
-    //     output.pipelineFunctions[`${name}`] = resolver.serialize();
-    //   }
+    const resolversBySource = new Map();
 
-    //   if (resolver instanceof FieldResolver) {
-    //     output.fieldResolvers[`${name}`] = resolver.serialize();
-    //   }
-    // }
+    for (const resolver of this.context.resolvers.values()) {
+      if (resolver.source) {
+        if (!resolversBySource.has(resolver.source)) {
+          resolversBySource.set(resolver.source, []);
+        }
+
+        resolversBySource.get(resolver.source).push(resolver);
+      }
+    }
+
+    const buildPaths = Array.from(resolversBySource.keys());
+
+    const build = buildSync({
+      entryPoints: buildPaths,
+      outdir: this.options.outputDirectory,
+      target: "esnext",
+      sourcemap: "inline",
+      sourcesContent: false,
+      treeShaking: true,
+      platform: "node",
+      format: "esm",
+      minify: false,
+      bundle: true,
+      write: false,
+      external: ["@aws-appsync/utils"],
+    });
+
+    if (build.errors.length) {
+      build.errors.forEach((error) => {
+        console.error(error);
+      });
+    }
+
+    for (const [index, sourcePath] of buildPaths.entries()) {
+      const resolverForPath = resolversBySource.get(sourcePath);
+      const file = build.outputFiles[Number(index)];
+
+      if (resolverForPath) {
+        for (const resolver of resolverForPath) {
+          if (resolver instanceof FunctionResolver) {
+            output.pipelineFunctions[`${resolver.name}`] = {
+              name: resolver.name,
+              dataSource: resolver.dataSource,
+              code: file.text,
+            };
+          }
+
+          if (resolver instanceof FieldResolver) {
+            output.fieldResolvers[`${resolver.typeName}.${resolver.fieldName}`] = {
+              typeName: resolver.typeName,
+              fieldName: resolver.fieldName,
+              pipelineFunctions: resolver.pipelineFunctions,
+              dataSource: resolver.dataSource,
+              code: file.text,
+            };
+          }
+        }
+      }
+    }
 
     return output;
   }
